+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 # Unreleased
 
+# v0.1.0-beta.6
+
+* feat: `icp identity export` to print the PEM file for the identity
+
 # v0.1.0-beta.5
 
 * fix: Fix error when loading network descriptors from v0.1.0-beta.3
 * feat: `icp identity delete` and `icp identity rename`
-* feat: `icp identity export` to print the PEM file for the identity
 
 # v0.1.0-beta.4